server/requirements.txt
fastapi[all]
sqlalchemy
alembic
psycopg[binary]
pydantic_settings
python-slugify
Markdown